+++ front/src/page/SignUp/SignUp.js
@@ -1,7 +1,9 @@
 import React, { useReducer, useCallback } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
-
+import InputEmail from './InputEmail';
+import InputPassword from './InputPassword';
+import InputConfirmPassword from './InputConfirmPassword';
 import * as S from './style';
 
 const initialState = {
@@ -31,68 +33,6 @@ const reducer = (state, action) => {
       return state;
   }
 };
-
-const InputEmail = React.memo(({ email, onChange, isValid }) => (
-  <>
-    <S.SignUpLabel htmlFor='email'>이메일</S.SignUpLabel>
-    <S.SignUpInput
-      type='email'
-      id='email'
-      placeholder='hong@gmail.com'
-      value={email}
-      onChange={onChange}
-      isvalid={isValid.toString()}
-      required
-    />
-    {!isValid && (
-      <S.SignUpValidationError>
-        올바른 이메일 주소를 입력해주세요.
-      </S.SignUpValidationError>
-    )}
-  </>
-));
-
-const InputPassword = React.memo(({ password, onChange, isValid }) => (
-  <>
-    <S.SignUpLabel htmlFor='password'>비밀번호</S.SignUpLabel>
-    <S.SignUpInput
-      type='password'
-      id='password'
-      placeholder='6자 이상, 숫자와 영문 필수 포함'
-      value={password}
-      onChange={onChange}
-      isvalid={isValid.toString()}
-      required
-    />
-    {!isValid && (
-      <S.SignUpValidationError>
-        비밀번호는 6자 이상이어야 하며, 숫자와 영문을 반드시 포함해야 합니다.
-      </S.SignUpValidationError>
-    )}
-  </>
-));
-
-const InputConfirmPassword = React.memo(
-  ({ confirmPassword, onChange, isValid }) => (
-    <>
-      <S.SignUpLabel htmlFor='confirmPassword'>비밀번호 확인</S.SignUpLabel>
-      <S.SignUpInput
-        type='password'
-        id='confirmPassword'
-        placeholder='비밀번호 확인'
-        value={confirmPassword}
-        onChange={onChange}
-        isvalid={isValid.toString()}
-        required
-      />
-      {!isValid && (
-        <S.SignUpValidationError>
-          비밀번호가 일치하지 않습니다.
-        </S.SignUpValidationError>
-      )}
-    </>
-  )
-);
 
 const SignUp = () => {
   const navigate = useNavigate();
@@ -138,7 +78,6 @@ const SignUp = () => {
     [state, navigate]
   );
 
-  // Input 값 변경 될 때마다 상태 없데이트 + 유효성 검사
   const handleEmailChange = useCallback((e) => {
     const email = e.target.value;
     dispatch({ type: 'SET_EMAIL', payload: email });
@@ -163,7 +102,6 @@ const SignUp = () => {
     [state.password]
   );
 
-  // 유효성 검사
   const validateEmail = (email) => {
     const emailRegex = /^[A-Za-z0-9]+@[A-Za-z0-9]+\.[A-Za-z]{2,}$/;
     return emailRegex.test(email);
@@ -182,7 +120,7 @@ const SignUp = () => {
         </S.GoBackButton>
         <S.SignUpTitle>회원정보 입력</S.SignUpTitle>
       </S.SignUpHeader>
-      <S.SignUpForm>
+      <S.SignUpForm onSubmit={handleSubmit}>
         <InputEmail
           email={state.email}
           onChange={handleEmailChange}
@@ -201,9 +139,7 @@ const SignUp = () => {
           isValid={state.isConfirmPasswordValid}
         />
 
-        <S.SignUpSubmitButton onClick={handleSubmit}>
-          Submit
-        </S.SignUpSubmitButton>
+        <S.SignUpSubmitButton type='submit'>Submit</S.SignUpSubmitButton>
       </S.SignUpForm>
     </>
   );
